requirements.txt.txt
google-generativeai
opencv-python
gtts
playsound==1.2.2